requirements.txt
ipykernel
praw
python-dotenv
pandas
matplotlib